requirements.txt
asgiref==3.6.0
Django==4.2
django-extensions==3.2.1
djangorestframework==3.14.0       
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
python-dotenv==0.21.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3